Add files via upload
This is the Colab notebook for the problem that we are trying to solve, which is classifying the image of the dog into its breeds. The full description of the problem and its evaluation can be found on Kaggle. From Dataset to Evaluation techniques and overview can be found there. I'm linking the kaggle link for the reference.
https://www.kaggle.com/c/dog-breed-identification/overview